+++ Mindscape.Raygun4Net.WebApi/Mindscape.Raygun4Net.WebApi.csproj
@@ -75,18 +75,22 @@
     <Compile Include="..\Mindscape.Raygun4Net4\WebClientHelper.cs">
       <Link>WebClientHelper.cs</Link>
     </Compile>
+    <Compile Include="Filters\RaygunWebApiActionFilter.cs" />
+    <Compile Include="Filters\RaygunWebApiExceptionFilter.cs" />
     <Compile Include="Properties\AssemblyVersionInfo.cs" />
     <Compile Include="Extensions\RaygunHttpContentExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Builders\RaygunWebApiRequestMessageBuilder.cs" />
     <Compile Include="IRaygunWebApiClientProvider.cs" />
     <Compile Include="RaygunWebApiClient.cs" />
+    <Compile Include="RaygunWebApiClientProvider.cs" />
     <Compile Include="RaygunWebApiDelegatingHandler.cs" />
-    <Compile Include="RaygunWebApiSelectors.cs" />
+    <Compile Include="RaygunWebApiHttpException.cs" />
     <Compile Include="RaygunWebApiControllerActivator.cs" />
     <Compile Include="RaygunWebApiExceptionLogger.cs" />
-    <Compile Include="RaygunWebApiFilters.cs" />
     <Compile Include="RaygunWebApiMessageBuilder.cs" />
+    <Compile Include="Selectors\RaygunWebApiActionSelector.cs" />
+    <Compile Include="Selectors\RaygunWebApiControllerSelector.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
