+++ Mindscape.Raygun4Net.WebApi/Selectors/RaygunWebApiActionSelector.cs
@@ -1,8 +1,6 @@
-﻿using System.Collections.Generic;
-using System.Net.Http;
+using System.Collections.Generic;
 using System.Web.Http;
 using System.Web.Http.Controllers;
-using System.Web.Http.Dispatcher;
 
 namespace Mindscape.Raygun4Net.WebApi
 {
@@ -36,41 +34,6 @@ namespace Mindscape.Raygun4Net.WebApi
     }
 
     internal IHttpActionSelector ConcreteSelector
-    {
-      get { return _concreteSelector; }
-    }
-  }
-
-  public class RaygunWebApiControllerSelector : IHttpControllerSelector
-  {
-    private readonly IHttpControllerSelector _concreteSelector;
-    private readonly IRaygunWebApiClientProvider _clientCreator;
-
-    internal RaygunWebApiControllerSelector(IHttpControllerSelector concreteSelector, IRaygunWebApiClientProvider clientCreator)
-    {
-      _concreteSelector = concreteSelector;
-      _clientCreator = clientCreator;
-    }
-
-    public HttpControllerDescriptor SelectController(HttpRequestMessage request)
-    {
-      try
-      {
-        return _concreteSelector.SelectController(request);
-      }
-      catch (HttpResponseException ex)
-      {
-        _clientCreator.GenerateRaygunWebApiClient(request).SendInBackground(ex, new List<string> { RaygunWebApiClient.UnhandledExceptionTag });
-        throw;
-      }
-    }
-
-    public System.Collections.Generic.IDictionary<string, HttpControllerDescriptor> GetControllerMapping()
-    {
-      return _concreteSelector.GetControllerMapping();
-    }
-
-    internal IHttpControllerSelector ConcreteSelector
     {
       get { return _concreteSelector; }
     }
